Persist the new ordinance in the city save file
When saving a city the ordinance simulator will query the ordinance to
see if it implements cIGZSerializable, which is used to serialize the
object for placement in the save game.
When loading a city the game will call our director's GetClassObject
method to load the serialized ordinance from the save game.

Added copy and move support to OrdinanceBase, OrdinancePropertyHolder
and cRZBaseString.
Added interface ids for a few of the game's classes and fixed a few
bugs in those headers.